+++ nix/modules/configuration/defaults.nix
@@ -49,6 +49,7 @@ in
           inherit (outputs.packages) lockCreateWrapper;
           inherit (outputs.packages) lockExtendWrapper;
           inherit (outputs.packages) installBuildMissingWrapper;
+          inherit (outputs.packages) createLockInstallWrapper;
           inherit (outputs.packages) allWrapper;
         }'';
     };
@@ -157,6 +158,7 @@ in
           inherit (config.outputs.packages) lockCreateWrapper;
           inherit (config.outputs.packages) lockExtendWrapper;
           inherit (config.outputs.packages) installBuildMissingWrapper;
+          inherit (config.outputs.packages) createLockInstallWrapper;
           inherit (config.outputs.packages) allWrapper;
         }
       );